requirements.txt
absl-py                   2.1.0
aiohappyeyeballs          2.4.6
aiohttp                   3.11.12
aiohttp-cors              0.7.0
aiosignal                 1.3.2
annotated-types           0.7.0
antlr4-python3-runtime    4.9.3
attrs                     25.1.0
boltons                   23.0.0
brotlipy                  0.7.0
cachetools                5.5.1
calmsize                  0.1.3
certifi                   2024.8.30
cffi                      1.15.1
charset-normalizer        2.0.4
clarabel                  0.10.0
click                     8.1.8
colorful                  0.5.6
conda                     23.9.0
conda-content-trust       0.2.0
conda-libmamba-solver     23.9.1
conda-package-handling    2.2.0
conda_package_streaming   0.9.0
contourpy                 1.3.0
cryptography              41.0.3
cvxpy                     1.5.4
cycler                    0.12.1
datasets                  3.1.0
dill                      0.3.8
distlib                   0.3.9
ecos                      2.0.14
faiss-cpu                 1.7.4
filelock                  3.16.1
flwr-datasets             0.4.0
fonttools                 4.54.1
frozenlist                1.5.0
fsspec                    2024.9.0
google-api-core           2.24.1
google-auth               2.38.0
googleapis-common-protos  1.67.0
grpcio                    1.70.0
h5py                      3.12.1
huggingface-hub           0.28.1
hydra-core                1.3.2
idna                      3.4
imageio                   2.36.1
Jinja2                    3.1.4
joblib                    1.4.2
jsonpatch                 1.32
jsonpointer               2.1
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
kiwisolver                1.4.7
lazy_loader               0.4
libmambapy                1.5.1
linkify-it-py             2.0.3
lpips                     0.1.4
Markdown                  3.7
markdown-it-py            3.0.0
MarkupSafe                3.0.2
matplotlib                3.9.2
mdit-py-plugins           0.4.2
mdurl                     0.1.2
memray                    1.15.0
mpmath                    1.3.0
msgpack                   1.1.0
multidict                 6.1.0
multiprocess              0.70.16
munch                     4.0.0
networkx                  3.4.2
numpy                     1.26.4
nvidia-cublas-cu12        12.1.3.1
nvidia-cuda-cupti-cu12    12.1.105
nvidia-cuda-nvrtc-cu12    12.1.105
nvidia-cuda-runtime-cu12  12.1.105
nvidia-cudnn-cu12         9.1.0.70
nvidia-cufft-cu12         11.0.2.54
nvidia-curand-cu12        10.3.2.106
nvidia-cusolver-cu12      11.4.5.107
nvidia-cusparse-cu12      12.1.0.106
nvidia-nccl-cu12          2.20.5
nvidia-nvjitlink-cu12     12.4.127
nvidia-nvtx-cu12          12.1.105
omegaconf                 2.3.0
opencensus                0.11.4
opencensus-context        0.1.3
opencv-python             4.10.0.84
osqp                      0.6.7.post3
packaging                 23.1
pandas                    2.2.3
patsy                     1.0.1
pillow                    10.3.0
pip                       23.2.1
platformdirs              4.3.6
pluggy                    1.0.0
pretrainedmodels          0.7.4
prometheus_client         0.21.1
propcache                 0.2.1
proto-plus                1.26.0
protobuf                  5.29.3
py-spy                    0.4.0
pyarrow                   16.1.0
pyasn1                    0.6.1
pyasn1_modules            0.4.1
pycosat                   0.6.6
pycparser                 2.21
pydantic                  2.10.6
pydantic_core             2.27.2
Pygments                  2.19.1
pynvml                    11.5.3
pyOpenSSL                 23.2.0
pyparsing                 3.2.0
PySocks                   1.7.1
python-dateutil           2.9.0.post0
pytorch-minimize          0.0.2
pytz                      2024.2
PyYAML                    6.0.2
qdldl                     0.1.7.post5
ray                       2.40.0
referencing               0.36.2
requests                  2.32.3
rich                      13.7.1
rpds-py                   0.22.3
rsa                       4.9
ruamel.yaml               0.17.21
safetensors               0.5.3
SciencePlots              2.1.1
scikit-image              0.24.0
scikit-learn              1.5.2
scipy                     1.13.1
scs                       3.2.7.post2
seaborn                   0.13.2
setuptools                68.0.0
six                       1.16.0
smart-open                7.1.0
statsmodels               0.14.4
sympy                     1.13.1
tensorboard               2.17.1
tensorboard-data-server   0.7.2
textual                   1.0.0
threadpoolctl             3.5.0
tifffile                  2024.9.20
timm                      1.0.15
torch                     2.4.1
torchvision               0.19.1
tornado                   6.4.2
tqdm                      4.67.1
triton                    3.0.0
truststore                0.8.0
tsnecuda                  3.0.1
typing_extensions         4.12.2
tzdata                    2024.2
uc-micro-py               1.0.3
ujson                     5.10.0
urllib3                   1.26.16
virtualenv                20.29.2
visdom                    0.2.4
websocket-client          1.8.0
Werkzeug                  3.1.3
wheel                     0.41.2
wrapt                     1.17.2
xlrd                      2.0.1
xxhash                    3.5.0
yarl                      1.18.3
zstandard                 0.19.0